refactor: move config
Move config to nerdlets/shared/config
