+++ nerdlets/mande-nerdlet/MandeContainer.js
@@ -19,9 +19,9 @@ import MetricSidebar from './components/metric-sidebar/MetricSidebar'
 import MetricDashboard from './components/dashboard/MetricDashboard'
 import CategoryDetail from './components/category-detail/CategoryDetail'
 import Selected from './components/metric-sidebar/Selected'
-import metricConfigs from '../../config/MetricConfig'
-import { FIND_USER_ATTRIBUTE } from '../../config/MetricConfig'
-import { activeEvents } from '../../config/VideoConfig'
+import metricConfigs from '../shared/config/MetricConfig'
+import { FIND_USER_ATTRIBUTE } from '../shared/config/MetricConfig'
+import { activeEvents } from '../shared/config/VideoConfig'
 import {
   formatFilters,
   formatFacets,
