+++ nerdlets/shared/config/AdConfig.js
