+++ nerdlets/shared/config/AwsCdnConfig.js
