+++ nerdlets/shared/config/AwsIngestConfig.js
