+++ nerdlets/shared/config/AwsOriginConfig.js
