+++ nerdlets/shared/config/AwsProcessConfig.js
