+++ nerdlets/shared/config/AwsPublishConfig.js
