+++ nerdlets/shared/config/AwsSourceConfig.js
