+++ nerdlets/shared/config/DeviceConfig.js
