+++ nerdlets/user-video-view/UserViewContainer.js
@@ -3,7 +3,7 @@ import { Stack, StackItem, HeadingText } from 'nr1'
 import { formatSinceAndCompare } from '../../utils/query-formatter'
 import { dateFormatter } from '../../utils/date-formatter'
 import QosKpiGrid from '../shared/components/qos/QosKpiGrid'
-import videoConfig from '../../config/VideoConfig'
+import videoConfig from '../shared/config/VideoConfig'
 import ViewTable from './components/view/ViewTable'
 
 export default class UserViewContainer extends React.Component {
